+++ sdk/core/azure-core/src/main/java/com/azure/core/implementation/serializer/HttpResponseDecodeData.java
@@ -9,9 +9,12 @@ import com.azure.core.http.rest.Response;
 import com.azure.core.http.rest.ResponseBase;
 import com.azure.core.implementation.TypeUtil;
 import com.azure.core.implementation.http.UnexpectedExceptionInformation;
+import com.azure.core.util.serializer.SerializerAdapter;
+import com.azure.core.util.serializer.SerializerEncoding;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
+import java.io.IOException;
 import java.lang.reflect.Type;
 
 /**
@@ -27,9 +30,9 @@ public interface HttpResponseDecodeData {
 
     /**
      * Get the type of the entity to be used to deserialize 'Matching' headers.
-     *
+     * <p>
      * The 'header entity' is optional and client can choose it when a strongly typed model is needed for headers.
-     *
+     * <p>
      * 'Matching' headers are the HTTP response headers those with:
      * 1. header names same as name of a properties in the 'header entity'.
      * 2. header names start with value of {@link HeaderCollection} annotation applied to the properties in the 'header
@@ -64,7 +67,7 @@ public interface HttpResponseDecodeData {
 
     /**
      * Get the type of the 'entity' in HTTP response content.
-     *
+     * <p>
      * When this method return non-null {@code java.lang.reflect.Type} then the raw HTTP response
      * content will need to parsed to this {@code java.lang.reflect.Type} then converted to actual
      * {@code returnType}.
@@ -123,4 +126,19 @@ public interface HttpResponseDecodeData {
      * @return Whether the network response body should be eagerly read.
      */
     boolean isResponseEagerlyRead();
+
+    /**
+     * Deserializes the response {@code data} into the {@code bodyType} using the specified {@code encoding}.
+     *
+     * @param serializer The {@link SerializerAdapter} used for deserialization.
+     * @param data The body.
+     * @param bodyType The body type.
+     * @param encoding The serialization encoding.
+     * @return The deserialized body.
+     * @throws IOException If an I/O error occurs during deserialization.
+     */
+    default Object deserializeBody(SerializerAdapter serializer, byte[] data, Type bodyType,
+        SerializerEncoding encoding) throws IOException {
+        return serializer.deserialize(data, bodyType, encoding);
+    }
 }